refactor(node/governance): mv maintenance mod to governance
