+++ opentron/src/manager/governance/mod.rs
@@ -1,3 +1,4 @@
 //! Chain governance related.
 
+pub mod maintenance;
 pub mod reward;
